# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse:
    properties:
      error:
        type: string
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.ValidationError'
        type: array
    type: object
...
  github_com_Aliizi83_vohu_internal_platform_shared.PagedList-internal_platform_rbac_ResourceAccessResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/internal_platform_rbac.ResourceAccessResponse'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  github_com_Aliizi83_vohu_internal_platform_shared.PagedList-internal_platform_rbac_RoleResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/internal_platform_rbac.RoleResponse'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_Aliizi83_vohu_internal_platform_shared.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 40901
    - 50001
    type: integer
    x-enum-varnames:
    - ResultSuccess
    - ResultValidationError
    - ResultAuthError
    - ResultForbiddenError
    - ResultNotFoundError
    - ResultConflictError
    - ResultInternalError
  github_com_Aliizi83_vohu_internal_platform_shared.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
  internal_platform_rbac.AccessLevel:
    enum:
    - read
    - write
    - manage
    type: string
    x-enum-varnames:
    - AccessRead
    - AccessWrite
    - AccessManage
  internal_platform_rbac.AssignRoleRequest:
    properties:
      roleId:
        type: integer
    required:
    - roleId
    type: object
  internal_platform_rbac.CreateRoleRequest:
    properties:
      name:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    type: object
  internal_platform_rbac.GrantResourceAccessRequest:
    properties:
      effect:
        allOf:
        - $ref: '#/definitions/internal_platform_rbac.ResourceEffect'
        enum:
        - accepted
        - prohibited
      granteeId:
        type: integer
      granteeType:
        allOf:
        - $ref: '#/definitions/internal_platform_rbac.GranteeType'
        enum:
        - user
        - role
      level:
        allOf:
        - $ref: '#/definitions/internal_platform_rbac.AccessLevel'
        enum:
        - read
        - write
        - manage
      resourceId:
        type: integer
      resourceType:
        maxLength: 50
        type: string
    required:
    - effect
    - granteeId
    - granteeType
    - level
    - resourceType
    type: object
  internal_platform_rbac.GranteeType:
    enum:
    - user
    - role
    type: string
    x-enum-varnames:
    - GranteeUser
    - GranteeRole
  internal_platform_rbac.MyAccessResponse:
    properties:
      levels:
        additionalProperties:
          $ref: '#/definitions/internal_platform_rbac.AccessLevel'
        type: object
      resourceAccess:
        items:
          $ref: '#/definitions/internal_platform_rbac.ResourceAccessResponse'
        type: array
    type: object
  internal_platform_rbac.ResourceAccessResponse:
    properties:
      effect:
        $ref: '#/definitions/internal_platform_rbac.ResourceEffect'
      granteeId:
        type: integer
      granteeType:
        $ref: '#/definitions/internal_platform_rbac.GranteeType'
      id:
        type: integer
      level:
        $ref: '#/definitions/internal_platform_rbac.AccessLevel'
      resourceId:
        type: integer
      resourceType:
        type: string
    type: object
  internal_platform_rbac.ResourceEffect:
    enum:
    - accepted
    - prohibited
    type: string
    x-enum-varnames:
    - EffectAccepted
    - EffectProhibited
  internal_platform_rbac.RoleResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  internal_platform_rbac.UpdateRoleRequest:
    properties:
      name:
        maxLength: 50
        minLength: 2
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API for Vohu's multi-user platform — users, roles, polymorphic
    resource access, SSH connections, conversations/chat, provider keys, custom models,
    and the agent tool catalog. Every endpoint responds through shared.BaseResponse
    (see internal/platform/shared/response.go) — {result, success, resultCode, validationErrors,
    error} — regardless of success or failure.
  title: Vohu Platform API
  version: "1.0"
paths:
...
  /me/access:
    get:
      description: Returns the caller's own complete access profile — every resource-access
        grant that's personally theirs (direct or via a role they hold), plus the
        best level they hold on each known resource type at large. The frontend uses
        this right after login to decide what nav items and buttons to show; it's
        a UX layer, every endpoint still enforces its own access independently regardless
        of what this reports.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/internal_platform_rbac.MyAccessResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get my access profile
      tags:
      - resource-access
...
  /resource-access:
    get:
      description: Lists every resource-access grant in the system. Requires wildcard
        "read" access on resource type "resource_access".
      parameters:
      - description: Page number, default 1
        in: query
        name: pageNumber
        type: integer
      - description: Page size, default 10
        in: query
        name: pageSize
        type: integer
      - description: JSON-encoded shared.DynamicFilter
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.PagedList-internal_platform_rbac_ResourceAccessResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: List resource-access grants
      tags:
      - resource-access
    post:
      consumes:
      - application/json
      description: Grants (or upserts, if one already exists for the same grantee+resource)
        a resource-access row. The caller must already hold "manage" on the target
        resource — granting access to X requires manage on X, no separate meta-permission.
      parameters:
      - description: Grant
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_platform_rbac.GrantResourceAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "403":
          description: Caller doesn't hold manage on the target resource
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Grant resource access
      tags:
      - resource-access
  /resource-access/{id}:
    delete:
      description: Deletes one resource-access grant by its own ID. The caller must
        hold "manage" on the resource that grant refers to.
      parameters:
      - description: Resource-access grant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Revoke a resource-access grant
      tags:
      - resource-access
  /roles:
    get:
      description: Lists every role. Requires wildcard "read" access on resource type
        "role".
      parameters:
      - description: Page number, default 1
        in: query
        name: pageNumber
        type: integer
      - description: Page size, default 10
        in: query
        name: pageSize
        type: integer
      - description: JSON-encoded shared.DynamicFilter
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.PagedList-internal_platform_rbac_RoleResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Creates a new role. Requires wildcard "write" access on resource
        type "role".
      parameters:
      - description: New role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_platform_rbac.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/internal_platform_rbac.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "409":
          description: A role with this name already exists
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Create a role
      tags:
      - roles
  /roles/{id}:
    delete:
      description: Deletes a role. Requires "manage" access to this specific role.
        Users holding the role are not deleted, just lose whatever the role granted
        them.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Delete a role
      tags:
      - roles
    get:
      description: Returns one role by ID. Requires at least "read" access to this
        specific role.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/internal_platform_rbac.RoleResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get a role
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Renames a role. Requires "write" access to this specific role.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: New name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_platform_rbac.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/internal_platform_rbac.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Update a role
      tags:
      - roles
...
  /users/{id}/roles:
    post:
      consumes:
      - application/json
      description: Grants a user every resource access attached to the given role.
        Requires "manage" access to the target user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role to assign
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_platform_rbac.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Assign a role to a user
      tags:
      - roles
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token from POST
      /auth/login or /auth/refresh.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"